% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serie_anime_color.R
\docType{data}
\name{palette_anime}
\alias{palette_anime}
\title{Complete list of palettes}
\format{
An object of class \code{list} of length 23.
}
\usage{
palette_anime
}
\description{
Use \code{names(palette_anime)} to view list of palette anime names.
}
\keyword{datasets}
