% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serie_netflyx_color.R
\docType{data}
\name{palette_netflyx}
\alias{palette_netflyx}
\title{Complete list of palettes}
\format{
An object of class \code{list} of length 6.
}
\usage{
palette_netflyx
}
\description{
Use \code{names(palette_netflyx)} to view list of palette netflyx names.
}
\keyword{datasets}
