% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serie_peru_color.R
\docType{data}
\name{palette_peru}
\alias{palette_peru}
\title{Complete list of palettes}
\format{
An object of class \code{list} of length 1.
}
\usage{
palette_peru
}
\description{
Use \code{names(palette_peru)} to view list of palette netflyx names.
}
\keyword{datasets}
