% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serie_palette.R
\name{Rseries_pal}
\alias{Rseries_pal}
\title{3. Return function to interpolate a Rseries color palette}
\usage{
Rseries_pal(
  palette = "Lupin",
  palette_family = palette_netflyx,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{Character name of palette in family Rserie palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
3. Return function to interpolate a Rseries color palette
}
