% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serie_palette.R
\name{scale_color_innova}
\alias{scale_color_innova}
\title{function to personalize fill aesthetic}
\usage{
scale_color_innova(
  palette_family = palette_netflyx,
  palette = "Lupin",
  discrete = TRUE,
  reverse = FALSE,
  ...
)
}
\description{
function to personalize fill aesthetic
}
