% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serie_palette.R
\name{scale_color_rseries}
\alias{scale_color_rseries}
\title{function to personalize color aesthetic}
\usage{
scale_color_rseries(
  palette_family = palette_netflyx,
  palette = "Lupin",
  discrete = TRUE,
  reverse = FALSE,
  ...
)
}
\description{
function to personalize color aesthetic
}
