% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serie_palette.R
\name{scale_fill_rseries}
\alias{scale_fill_rseries}
\title{function to personalize fill aesthetic}
\usage{
scale_fill_rseries(
  palette_family = palette_netflyx,
  palette = Lupin,
  discrete = TRUE,
  reverse = FALSE,
  ...
)
}
\description{
function to personalize fill aesthetic
}
