% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serie_palette.R
\name{show_pal}
\alias{show_pal}
\title{function to show all the avaliable palettes in a specific family .}
\usage{
show_pal(
  name = "all",
  palette_family = palette_netflyx,
  n = 6,
  rev = TRUE,
  ...
)
}
\description{
function to show all the avaliable palettes in a specific family .
}
