% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serie_palette.R
\name{Rseries_pal}
\alias{Rseries_pal}
\title{3. Return function to interpolate a Rseries color palette
 palette color intepolate}
\usage{
Rseries_pal(
  palette = "Lupin",
  palette_family = palette_netflyx,
  reverse = FALSE,
  ...
)
}
\description{
3. Return function to interpolate a Rseries color palette
 palette color intepolate
}
